fix : CAP-38 - refactoring the duplicated code
