--- src/Pages/FixturePage.css
@@ -1,10 +1,0 @@
-.fixture-page-container {
-    margin-top: 100px;
-    width: 75%; /* Set the width to 75% of the container */
-    margin-left: 10px;
-}
-
-.fixture-table-container {
-    margin-left: 20px; /* Automatically adjust margin on the left to push the table to the right */
-    margin-top: 20px; /* Add more space between the heading and the table */
-}
